// packages/ui/src/components/radio/radio.types.ts
/**
 * Copyright 2023 Arkemis S.r.l.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { ChangeEventHandler, ReactNode } from "react";

interface RadioProps {
  name?: string;
  label?: string | ReactNode;
  labelPlacement?: "top" | "bottom" | "left" | "right";
  color?: "primary" | "secondary";
  checked?: boolean;
  onChange?: ChangeEventHandler<HTMLInputElement>;
  value?: string;
  disabled?: boolean;
}

export type { RadioProps };